The 1D CNN network training is completed with early stopping and PR-AUC search.
Without scaling (RobustScaler), the best PR-AUC=0.7238 and ROC-AUC=0.4929 are obtained in validation, and PR-AUC=0.7575 and ROC-AUC=0.5215 in testing.

With scaling, the best PR-AUC=0.8346 and ROC-AUC=0.6509 are obtained in validation, and PR-AUC=0.8262 and ROC-AUC=0.61897 in testing.

Hyperparameter optimization with scaled data, early stopping, and PR-AUC search yields the best PR. The final 1D CNN network is trained with better hyperparameters, obtaining a better PR-AUC=0.8380  and a ROC-AUC=0.6593 in validation, and a PR-AUC=0.8366  and a ROC-AUC=0.6393 in testing.